Implement better Creeper mobGriefing game rule
Add new "mobGriefingCreeper" game rule, this will be a boolean value and
will override the default "mobGriefing" game rule when the entity is a
Creeper.

Add framework methods to BetterMobGriefingGameRule to add the new
game rules to a world and determine which new game rule should be used
based on an entity type.
Add the new Creeper game rule to the framework.

Create BetterMobGriefingGameRuleEventHandler, handle the "Detonate"
event to control block damage for Creeper explosions.

Close #1