Implement better Wither mobGriefing game rule
Add new "mobGriefingWither" game rule, this will be a boolean value and
will override the default "mobGriefing" game rule when the entity is a
Wither.

Modify the event handler's onDetonateEvent to support explosions from
WitherSkulls. The exploder field is populated with the WitherSkull and
the shooting entity is populated and retrieved directly.
The on spawn explosion is handled with existing onDetonateEvent
handling.

Handle the Wither's destruction of blocks around itself when hurt, which
can not be handled within existing events.
Add a new AI task which will be registered to Wither's on spawn, this
task will perform two actions:

- When mobGriefing is overridden as true the block destruction code will
run and destroy blocks around the Wither.
- When mobGriefing is overridden as false the block destruction cooldown
will be reset so that the original handling never triggers.

The task will avoid the need to perform ASM bytecode manipulation to
override the default mobGriefing rule for EntityWither.

Close #7