Implement better Silverfish mobGriefing game rule
Add new "mobGriefingSilverfish" game rule, this will be a boolean value
and will override the default "mobGriefing" game rule when the entity is
a Silverfish.

Add a method to handle LivingUpdateEvent to the event handler, this will
override the onLivingUpdate mob griefing handling done in the
EntitySilverfish class.
Silverfish mob griefing occurs when an ally is summoned from a monster
egg, when mob griefing is enabled the monster egg block is destroyed but
if mob griefing is disabled the block is replaced by its stone
equivalent.

The EntitySilverfish mob griefing will be controlled by copying the
summon ally code and running it with the desired mobGriefing rule during
the event.

Disabling the original handling will be done by manipulating the
"allySummonCooldown" field in the EntitySilverfish class.
The LivingUpdateEvent handler will disable the original handling of ally
summoning by ensuring the cooldown field is always set to -1.
The actual value of the cooldown will be stored in the entity data of
the Silverfish.

Close #6