Implement better EnderDragon mobGriefing game rule
Add new "mobGriefingDragon" game rule, this will be a boolean value
and will override the default "mobGriefing" game rule when the entity is
an EntityDragon.

Add EntityDragon to the classes handled by the mobGriefing ASM
transformer. This will replace instance of the "mobGriefing" game rule
in the EntityDragon class with "mobGriefingDragon".

Close #2