Remove hard coded mobGriefing game rule strings
The mobGriefing rules for each entity are stored as hard coded strings,
when determining which rule to use the entity type is checked and the
matching string is returned.
This causes a lot of overhead whenever new entities need to be added and
stops potential usage of custom entities or custom behaviours using the
introduced entity specific rules.

Remove the hard coding of the string constants, instead the EntityList
can be used to get the registered entities name and the rule built
dynamically.

Update BetterMobGriefingGameRule.getMobGriefingRule(), remove the entity
class instance checks and replace it with a query of EntityList and
concatenation to build the entity specific rule dynamically.

Remove the entity specific tests from BetterMobGriefingGameRuleTest and
create new tests which just test the existence and non-existence of an
entity specific rule. It is no longer required to test multiple entity
types.

Close #10