Update README.md to reflect game rule changes
Update the README.md to show the new usage of the better "mobGriefing"
game rule.
The entity specific game rules are now an extension of the original rule
in the form "/gamerule mobGriefing <entity name> <value>".
Include the usage of the "inherit" value to allow entity behaviour to
follow changes in the global game rule.