Implement support for loot pick up mob griefing
EntityZombie and EntitySkeleton based entities are able to pick up dropped
equipment based on a "canPickUpLoot" variable being set as well as the
value of the mobGriefing game rule.
Update BetterMobGriefingGameRuleIClassTransformer to add a transform
target for EntityLiving.onLivingUpdate, the existing transformation will
be able to perform the transformation with just updated targets.