Improve Silverfish mob griefing maintainability
The current implementation to handle Silverfish mob griefing uses the
onLivingUpdateEvent and overrides a timer to stop block damage. However,
to maintain the existing behaviour a large chunk of Silverfish code was
copied to the event handler and was ran when needed to force the
behaviour.
When moving to future versions of Minecraft this approach requires
disproportionate amount of work due to the duplicated code having to be
checked for changes in the later iterations.
To resolve this the current handling should be removed and a Silverfish
target should be added to the core mod transformer. In this case it is
seen as appropriate to use the core mod due to the maintainability
concerns.

The entity index used in the EntityDragon transform target is incorrect
and should be zero, update the entity index to the correct value.